api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
google-generativeai==0.3.2
python-dotenv==1.0.1
pytest==8.0.0
httpx==0.26.0
loguru==0.7.2
openai==1.75.0